wip(#369): split 8 middle blocks (word-nav/dropdown-render/caret/dropdown-apply/input-mode/queue/paste/picker); replicate clipboard vi.mock per-file
