+++ src/cli/terminal-compositor.test.ts
@@ -2635,2627 +2635,3 @@ describe('TerminalCompositor', () => {
 // existing `left/right move cursor within buffer` test on line ~983.
 // ---------------------------------------------------------------------------
 
-describe('TerminalCompositor — word/line navigation (readline parity)', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-  });
-
-  // Helper: type each char of `text` so the buffer reaches `text` with
-  // cursor at end. Mirrors the typing pattern used throughout this file.
-  const type = (s: string) => {
-    for (const ch of s) {
-      stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    }
-  };
-
-  describe('cursor movement', () => {
-    it('Ctrl+A moves to start of line (Cmd+← via terminal default remap)', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello');
-      stdin.emit('keypress', undefined, { name: 'a', ctrl: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('Xhello');
-    });
-
-    it('Ctrl+E moves to end of line (Cmd+→ via terminal default remap)', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello');
-      // Move to start, then Ctrl+E to jump back to end.
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', undefined, { name: 'e', ctrl: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('helloX');
-    });
-
-    it('Option+← (meta+left) moves backward one word', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('foo bar baz');
-      // Cursor at end (11). Option+← lands at start of 'baz' (cursor=8).
-      stdin.emit('keypress', undefined, { name: 'left', meta: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('foo bar Xbaz');
-    });
-
-    it('Option+→ (meta+right) moves forward one word', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('foo bar baz');
-      stdin.emit('keypress', undefined, { name: 'home' });
-      // Cursor at 0. Option+→ lands at end of 'foo' (cursor=3).
-      stdin.emit('keypress', undefined, { name: 'right', meta: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('fooX bar baz');
-    });
-
-    it('Ctrl+← (ctrl+left) moves backward one word (Linux convention)', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('alpha beta');
-      stdin.emit('keypress', undefined, { name: 'left', ctrl: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('alpha Xbeta');
-    });
-
-    it('Ctrl+→ (ctrl+right) moves forward one word', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('alpha beta');
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', undefined, { name: 'right', ctrl: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('alphaX beta');
-    });
-
-    it('Option+B (Esc-prefixed, meta+b) moves backward one word', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('one two three');
-      stdin.emit('keypress', 'b', { name: 'b', meta: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('one two Xthree');
-    });
-
-    it('Option+F (Esc-prefixed, meta+f) moves forward one word', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('one two three');
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', 'f', { name: 'f', meta: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('oneX two three');
-    });
-
-    it('word-nav across whitespace boundary skips the gap', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('a   b');
-      // Cursor at 5 (end). One Option+← jumps over the spaces AND over 'b'
-      // to land at start of 'b' (cursor=4). Match InputCore.moveWordBackward
-      // semantics: skip trailing whitespace, then preceding non-whitespace run.
-      stdin.emit('keypress', undefined, { name: 'left', meta: true });
-      stdin.emit('keypress', 'X', { name: 'X', sequence: 'X' });
-      expect(c.getBuffer().text).toBe('a   Xb');
-    });
-  });
-
-  describe('word-level delete', () => {
-    it('Ctrl+W deletes word backward', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello world');
-      stdin.emit('keypress', undefined, { name: 'w', ctrl: true });
-      expect(c.getBuffer().text).toBe('hello ');
-    });
-
-    it('Option+Delete (meta+backspace) deletes word backward', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello world');
-      stdin.emit('keypress', undefined, { name: 'backspace', meta: true });
-      expect(c.getBuffer().text).toBe('hello ');
-    });
-
-    it('Option+Fn-Delete (meta+delete) deletes word forward', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello world');
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', undefined, { name: 'delete', meta: true });
-      expect(c.getBuffer().text).toBe(' world');
-    });
-
-    it('plain backspace still does char-erase (meta-variant does not regress unmodified key)', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('abc');
-      stdin.emit('keypress', undefined, { name: 'backspace' });
-      expect(c.getBuffer().text).toBe('ab');
-    });
-
-    it('plain delete still does forward char-erase', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('abc');
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', undefined, { name: 'delete' });
-      expect(c.getBuffer().text).toBe('bc');
-    });
-  });
-
-  describe('line-level delete', () => {
-    it('Ctrl+U deletes from cursor to start of line', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello world');
-      // Cursor at 11 (end). Ctrl+U kills the entire line back to start.
-      stdin.emit('keypress', undefined, { name: 'u', ctrl: true });
-      expect(c.getBuffer().text).toBe('');
-    });
-
-    it('Ctrl+K deletes from cursor to end of line', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('hello world');
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', undefined, { name: 'k', ctrl: true });
-      expect(c.getBuffer().text).toBe('');
-    });
-
-    it('Ctrl+U mid-buffer deletes only the prefix before the cursor', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      type('abcdef');
-      // Move left 3 times → cursor=3 (between c and d). Ctrl+U deletes 'abc'.
-      stdin.emit('keypress', undefined, { name: 'left' });
-      stdin.emit('keypress', undefined, { name: 'left' });
-      stdin.emit('keypress', undefined, { name: 'left' });
-      stdin.emit('keypress', undefined, { name: 'u', ctrl: true });
-      expect(c.getBuffer().text).toBe('def');
-    });
-  });
-
-  describe('history.resetRecall coordination', () => {
-    // Per the existing convention (e.g. backspace at line 1591), any
-    // buffer-modifying op must call history.resetRecall() so the next
-    // ↑/↓ recall starts from the edited buffer rather than the prior
-    // recalled snapshot. Pure cursor moves must NOT reset recall —
-    // verified by omission below (no test asserts that Ctrl+A/E/word-nav
-    // resets recall, because they shouldn't).
-
-    it('Ctrl+W resets history recall', async () => {
-      const resetRecall = vi.fn();
-      const history = {
-        push: vi.fn(),
-        back: vi.fn(() => null),
-        forward: vi.fn(() => null),
-        resetRecall,
-        get inRecall() { return false; },
-      };
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), history });
-      await c.arm();
-      type('hello world');
-      resetRecall.mockClear();
-      stdin.emit('keypress', undefined, { name: 'w', ctrl: true });
-      expect(resetRecall).toHaveBeenCalledTimes(1);
-      expect(c.getBuffer().text).toBe('hello ');
-    });
-
-    it('Ctrl+U resets history recall', async () => {
-      const resetRecall = vi.fn();
-      const history = {
-        push: vi.fn(),
-        back: vi.fn(() => null),
-        forward: vi.fn(() => null),
-        resetRecall,
-        get inRecall() { return false; },
-      };
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), history });
-      await c.arm();
-      type('abc');
-      resetRecall.mockClear();
-      stdin.emit('keypress', undefined, { name: 'u', ctrl: true });
-      expect(resetRecall).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe('no-op safety at edges', () => {
-    it('Option+← at buffer start is a silent no-op', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // Empty buffer, cursor at 0.
-      stdin.emit('keypress', undefined, { name: 'left', meta: true });
-      // No throw, no buffer change.
-      expect(c.getBuffer().text).toBe('');
-    });
-
-    it('Ctrl+W on empty buffer is a silent no-op (does not reset recall)', async () => {
-      const resetRecall = vi.fn();
-      const history = {
-        push: vi.fn(),
-        back: vi.fn(() => null),
-        forward: vi.fn(() => null),
-        resetRecall,
-        get inRecall() { return false; },
-      };
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), history });
-      await c.arm();
-      resetRecall.mockClear();
-      stdin.emit('keypress', undefined, { name: 'w', ctrl: true });
-      // deleteWordBackward returns input unchanged → resetRecall not called.
-      expect(resetRecall).not.toHaveBeenCalled();
-      expect(c.getBuffer().text).toBe('');
-    });
-
-    it('Ctrl+B (background-turn binding) is preserved and does NOT trigger char-nav', async () => {
-      // Regression guard: AFK repurposes Ctrl+B for "background current
-      // turn" — see HOT context. Adding word-nav bindings must not
-      // accidentally bind Ctrl+B to char-back. Default input mode is
-      // `'streaming'`, where Ctrl+B fires onBackground.
-      const onBackground = vi.fn();
-      const c = new TerminalCompositor({
-        stdout,
-        stdin,
-        onCancel: vi.fn(),
-        onBackground,
-      });
-      await c.arm();
-      type('abc');
-      stdin.emit('keypress', undefined, { name: 'b', ctrl: true });
-      // Buffer must be untouched (no implicit cursor-back).
-      expect(c.getBuffer().text).toBe('abc');
-      expect(onBackground).toHaveBeenCalledTimes(1);
-    });
-  });
-});
-
-// ---------------------------------------------------------------------------
-// renderDropdownRows() — byte-level coverage: candidate text in stdout frame
-// ---------------------------------------------------------------------------
-
-describe('TerminalCompositor — renderDropdownRows() output', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-    // Seed the slash registry so filterSlashCandidates returns a real candidate
-    // when the buffer contains '/'. Isolated per-test via afterEach cleanup.
-    resetSlashRegistry();
-    registerSlashCommand({
-      name: '/render-test',
-      summary: 'Stub for renderDropdownRows coverage',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-  });
-
-  afterEach(() => {
-    resetSlashRegistry();
-  });
-
-  it('candidate text appears in the stdout frame when dropdownOpen is true', async () => {
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-
-    writes.clear();
-
-    // Type '/' to trigger slash autocomplete — updateAutocomplete() will call
-    // filterSlashCandidates('') which finds '/render-test' and sets dropdownOpen = true.
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-
-    // Dropdown must be open (earned via keystroke, not mutation).
-    expect(ac.dropdownOpen).toBe(true);
-
-    // The rendered frame written to stdout must contain the candidate value.
-    const frame = writes.all();
-    expect(frame).toContain('/render-test');
-  });
-
-  it('paste recomputes autocomplete and closes a now-stale slash dropdown (PR #574 regression)', async () => {
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-
-    // readClipboardImage is module-mocked; resolve it so the post-paste
-    // clipboard probe does not reject on an undefined return.
-    const clip = await import('./input/clipboard-image.js');
-    (clip.readClipboardImage as unknown as ReturnType<typeof vi.fn>).mockResolvedValue(null);
-
-    // Open the slash dropdown by typing '/'.
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-
-    // Paste a 6-line blob (>= 5 newlines -> truncates to a placeholder). The
-    // applyEdit guard skips updateAutocomplete during the paste burst and
-    // maybeTruncatePaste mutates the buffer directly, so the paste-end path
-    // MUST call updateAutocomplete or the dropdown renders stale. Regression
-    // guard for the PR #574 stale-dropdown fix.
-    stdin.emit('keypress', undefined, { sequence: '\x1b[200~' });
-    for (const ch of 'a\nb\nc\nd\ne\nf') {
-      if (ch === '\n') stdin.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-      else stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    }
-    stdin.emit('keypress', undefined, { sequence: '\x1b[201~' });
-    await new Promise((r) => setImmediate(r));
-
-    // The buffer no longer holds a bare slash token at the cursor, so the
-    // dropdown must have been recomputed and closed.
-    expect(ac.dropdownOpen).toBe(false);
-  });
-
-  it('input line is rendered AFTER dropdown rows in the frame (input pinned at bottom)', async () => {
-    // Invariant: the compositor frame's last line is always the input row,
-    // so log-update + DECSTBM pin the input one row above the status line
-    // regardless of dropdown / hint / overlay state. The dropdown grows
-    // UPWARD from the input — opening it does not shove the input up.
-    //
-    // We verify by typing `/` to open the dropdown, then asserting the
-    // dropdown candidate text appears in the captured frame BEFORE the
-    // typed `/` character that lives on the input row. log-update writes
-    // the joined frame top-to-bottom in one string, so byte order of the
-    // first occurrence reflects vertical row order.
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(), autocompleteState: ac, promptText: '> ',
-    });
-    await c.arm();
-
-    writes.clear();
-
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-
-    const frame = writes.all();
-    // The dropdown's `/render-test` row precedes the input row's
-    // prompt+slash by frame-build order. The input row is identifiable
-    // by the prompt prefix `> ` immediately followed by the `/` the
-    // user just typed.
-    const candidateIdx = frame.lastIndexOf('/render-test');
-    const inputRowIdx = frame.lastIndexOf('> ');
-    expect(candidateIdx).toBeGreaterThanOrEqual(0);
-    expect(inputRowIdx).toBeGreaterThanOrEqual(0);
-    // Candidate must appear earlier in the joined frame string than the
-    // input row marker — meaning the candidate is rendered ABOVE the
-    // input visually. Before the fix the order was reversed.
-    expect(candidateIdx).toBeLessThan(inputRowIdx);
-  });
-
-  it('hint tooltip (↳ <when-to-use>) renders between dropdown and input when selected candidate has a hint', async () => {
-    // The hint row is the `formatHintRow` tooltip — a `↳ <hint>` line
-    // shown for the currently-highlighted candidate. In the bottom-pinned
-    // layout it sits BELOW the dropdown rows and DIRECTLY ABOVE the input,
-    // so the tooltip for the selected candidate is adjacent to the
-    // cursor — same visual relationship the legacy reader.ts achieved
-    // when the dropdown lived below the input.
-    resetSlashRegistry();
-    registerSlashCommand({
-      name: '/hinted-cmd',
-      summary: 'one-line summary',
-      hint: 'When you need the long-form when-to-use tooltip',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(), autocompleteState: ac, promptText: '> ',
-    });
-    await c.arm();
-    writes.clear();
-
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-
-    const frame = writes.all();
-    // The `↳` glyph is the structural marker for the hint row (palette.dim
-    // wraps it but the bare character survives ANSI). Locate the first
-    // occurrence and check the relative positioning against the dropdown
-    // row above and the input row below.
-    const candidateIdx = frame.lastIndexOf('/hinted-cmd');
-    const hintIdx = frame.indexOf('↳');
-    const inputRowIdx = frame.lastIndexOf('> ');
-    expect(candidateIdx).toBeGreaterThanOrEqual(0);
-    expect(hintIdx).toBeGreaterThanOrEqual(0);
-    expect(inputRowIdx).toBeGreaterThanOrEqual(0);
-    // Frame order top-to-bottom: dropdown → hint → input.
-    expect(candidateIdx).toBeLessThan(hintIdx);
-    expect(hintIdx).toBeLessThan(inputRowIdx);
-  });
-
-  it('no `↳` glyph in the frame when the selected candidate has no hint (reserved slot only)', async () => {
-    // The default `/render-test` registered in the suite's beforeEach has
-    // no `hint` field — `formatHintRow` returns null. The frame still
-    // reserves a blank row in that slot (so the dropdown above doesn't
-    // shift when the user navigates onto a hinted candidate), but no
-    // visible tooltip glyph should appear.
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-    writes.clear();
-
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-
-    expect(writes.all()).not.toContain('↳');
-  });
-
-  it('frame row count is stable when navigating between hinted and un-hinted candidates', async () => {
-    // Regression for PR #478: a previous version of renderHintRow()
-    // returned null when the selected candidate had no `hint`, so the
-    // frame oscillated between N and N+1 rows as the user navigated ↑/↓
-    // across a hinted ↔ un-hinted boundary. The input row is pinned at
-    // `rows-1` via CupFrameRenderer, so the dropdown above it visibly
-    // shifted up by one row each navigation step — perceived as "the
-    // compositor jumping up and down" while scrolling the menu.
-    //
-    // Many real commands carry no hint (`/allow-dir`, `/bgsub`,
-    // `/changelog`, `/keys`, `/stats`, `/worktree`, the `/bgsub:*`
-    // variants), so this boundary crossing fires constantly in
-    // practice. Fix: always reserve a hint-row slot when the dropdown
-    // is open, even if it renders blank.
-    //
-    // We assert by spying on CupFrameRenderer.prototype.render and
-    // capturing the exact frame string passed on each call. The first
-    // argument is `frameLines.join('\n')`, so counting `\n` in that
-    // string is the literal row-count proxy. (We can't count `\n` in
-    // the captured stdout writes — the renderer emits CUP escapes
-    // instead of bare `\n`, so a stdout-based newline count would
-    // trivially pass with 0 === 0 and miss the regression entirely.)
-    const renderSpy = vi.spyOn(CupFrameRenderer.prototype, 'render');
-
-    resetSlashRegistry();
-    registerSlashCommand({
-      name: '/aaa-hinted',
-      summary: 'first alphabetically, hinted',
-      hint: 'When you want the hint to render',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-    registerSlashCommand({
-      name: '/zzz-unhinted',
-      summary: 'second alphabetically, no hint',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(), autocompleteState: ac, promptText: '> ',
-    });
-    await c.arm();
-
-    // First repaint: open dropdown. selectedIndex = 0 → /aaa-hinted (has hint).
-    renderSpy.mockClear();
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-    expect(ac.candidates[ac.selectedIndex]?.value).toBe('/aaa-hinted');
-    // The keypress can trigger more than one repaint (buffer edit +
-    // dropdown-open transition). The LAST render call carries the
-    // final settled frame for this keystroke — that's the one whose
-    // row count must match the post-navigation settled frame.
-    const hintedCalls = renderSpy.mock.calls;
-    expect(hintedCalls.length).toBeGreaterThan(0);
-    const hintedFrameStr = String(hintedCalls[hintedCalls.length - 1]![0]);
-    const rowsHinted = hintedFrameStr.split('\n').length;
-
-    // Second repaint: advance to /zzz-unhinted (index 1, no hint). The
-    // dropdown renders REVERSED (index 0 pinned at the bottom, growing
-    // upward), so moving to a HIGHER index is ↑, not ↓ — see the geometry
-    // Invariant in handleVerticalNav().
-    renderSpy.mockClear();
-    stdin.emit('keypress', '', { name: 'up', sequence: '\x1b[A' });
-    expect(ac.candidates[ac.selectedIndex]?.value).toBe('/zzz-unhinted');
-    const unhintedCalls = renderSpy.mock.calls;
-    expect(unhintedCalls.length).toBeGreaterThan(0);
-    const unhintedFrameStr = String(unhintedCalls[unhintedCalls.length - 1]![0]);
-    const rowsUnhinted = unhintedFrameStr.split('\n').length;
-
-    // Both frames must have the same row count — the un-hinted frame
-    // reserves a blank slot where the hinted frame draws the `↳ …`
-    // tooltip. Without the reservation, the un-hinted frame would be
-    // exactly one row shorter and the compositor would visually jump.
-    // Non-zero guard so a future regression that bypasses the render
-    // path (and captures zero frames) can't trivially pass.
-    expect(rowsHinted).toBeGreaterThan(1);
-    expect(rowsUnhinted).toBe(rowsHinted);
-
-    // Sanity: confirm exactly one of the two captured frames carries
-    // the `↳` glyph (so the test is actually exercising the hint /
-    // no-hint boundary, not a false-negative where neither frame has
-    // a hint).
-    expect(hintedFrameStr).toContain('↳');
-    expect(unhintedFrameStr).not.toContain('↳');
-
-    renderSpy.mockRestore();
-  });
-
-  it('selected dropdown candidate is rendered closest to the input (last among dropdown rows)', async () => {
-    // Fish/zsh-style invariant: when the dropdown grows upward, the
-    // candidate at viewportStart (the selected-by-default index 0)
-    // appears at the BOTTOM of the dropdown block — adjacent to the
-    // input row. Higher candidate indices ascend visually away from the
-    // input. Verified by registering a second slash command (so the
-    // dropdown has at least two visible rows) and asserting the
-    // alphabetically-earlier `/aaa-stub` (index 0, selected) appears
-    // AFTER the index-1 candidate in the frame's joined byte stream.
-    registerSlashCommand({
-      name: '/aaa-stub',
-      summary: 'first by sort order',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-    registerSlashCommand({
-      name: '/zzz-stub',
-      summary: 'last by sort order',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(), autocompleteState: ac, promptText: '> ',
-    });
-    await c.arm();
-    writes.clear();
-
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-    // selectedIndex starts at 0; candidates are alphabetized by filterSlashCandidates.
-    expect(ac.candidates[ac.selectedIndex]?.value).toBe('/aaa-stub');
-
-    const frame = writes.all();
-    const zzzIdx = frame.lastIndexOf('/zzz-stub');
-    const aaaIdx = frame.lastIndexOf('/aaa-stub');
-    expect(zzzIdx).toBeGreaterThanOrEqual(0);
-    expect(aaaIdx).toBeGreaterThanOrEqual(0);
-    // /zzz-stub (index 1) is rendered ABOVE /aaa-stub (selected, index 0),
-    // so it appears earlier in the joined frame string.
-    expect(zzzIdx).toBeLessThan(aaaIdx);
-  });
-
-  it('no candidate text in stdout frame when dropdownOpen is false', async () => {
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-
-    writes.clear();
-
-    // Type a non-trigger character — no dropdown.
-    stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-
-    expect(ac.dropdownOpen).toBe(false);
-    expect(writes.all()).not.toContain('/render-test');
-  });
-
-  it('delete key resets history recall (mirrors reader.ts delete behaviour)', async () => {
-    // Regression: compositor delete branch must call history.resetRecall()
-    // so _draft/_index are not corrupted when the user presses Delete while
-    // in history-recall mode inside the compositor.
-    const { makeHistory } = await import('./input/autocomplete-state.test.js').catch(
-      () => ({ makeHistory: null }),
-    );
-    // makeHistory is local to autocomplete-state.test.ts — define a minimal
-    // inline version here to avoid cross-test coupling.
-    const resetRecall = vi.fn();
-    const history = {
-      back: vi.fn(() => 'entry'),
-      forward: vi.fn(() => null),
-      resetRecall,
-      get inRecall() { return false; },
-    };
-
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), history });
-    await c.arm();
-
-    // Pre-load buffer with text so delete has something to remove.
-    stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-    stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-    // Move cursor to start so delete acts on the first char.
-    stdin.emit('keypress', undefined, { name: 'home' });
-
-    resetRecall.mockClear();
-
-    // Delete forward — must call history.resetRecall().
-    stdin.emit('keypress', undefined, { name: 'delete' });
-
-    expect(resetRecall).toHaveBeenCalledTimes(1);
-    // Buffer should now be 'b' (deleted 'a').
-    expect(c.getBuffer().text).toBe('b');
-  });
-});
-
-// ---------------------------------------------------------------------------
-// formatInputBuffer callback (Fix A): the compositor passes pre-cursor /
-// post-cursor segments through a caller-supplied formatter so colorization
-// can be wired without coupling the compositor to the slash registry.
-// ---------------------------------------------------------------------------
-
-describe('TerminalCompositor — formatInputBuffer callback', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-  });
-
-  it('invokes formatInputBuffer with pre-cursor and post-cursor segments separately', async () => {
-    const calls: string[] = [];
-    const formatInputBuffer = vi.fn((segment: string) => {
-      calls.push(segment);
-      return `[${segment}]`;
-    });
-
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), formatInputBuffer });
-    await c.arm();
-
-    // Type 'abc' (cursor=3, past end → cursorText=' ', rawBefore='abc', rawAfter='')
-    stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-    stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-    stdin.emit('keypress', 'c', { name: 'c', sequence: 'c' });
-    // Two left arrows → cursor=1 on 'b'. rawBefore='a', rawAfter='c'.
-    // (Cursor text 'b' is rendered raw via inverse SGR — does NOT pass through
-    // the formatter; that's covered by the dedicated test below.)
-    stdin.emit('keypress', undefined, { name: 'left' });
-    stdin.emit('keypress', undefined, { name: 'left' });
-
-    // The formatter must have been called with both segment shapes by the
-    // final render — 'a' as the pre-cursor segment and 'c' as the post-cursor
-    // segment after the second left arrow.
-    expect(calls).toContain('a');
-    expect(calls).toContain('c');
-    // Rendered frame must contain the formatter's bracket wrappers, proving
-    // formatter output (not raw segments) is what reached the render path.
-    expect(writes.all()).toContain('[a]');
-    expect(writes.all()).toContain('[c]');
-  });
-
-  it('renders raw segments when formatInputBuffer is not provided (backward-compat)', async () => {
-    // Sanity: pre-existing callers that did not pass a formatter must keep working.
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    writes.clear();
-    stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-
-    // No throw, frame includes the raw typed char.
-    expect(writes.all()).toContain('x');
-  });
-
-  it('does NOT pass the cursor character through the formatter', async () => {
-    // The inverse-video cursor block is rendered raw so chained ANSI codes
-    // (inverse + colorizer SGRs) don't compose into a broken cell. Verify by
-    // checking the cursor char never appears as a formatter argument.
-    const seen: string[] = [];
-    const formatInputBuffer = (segment: string) => {
-      seen.push(segment);
-      return segment;
-    };
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), formatInputBuffer });
-    await c.arm();
-    stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-    stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-    stdin.emit('keypress', undefined, { name: 'left' });
-    // Now buffer='ab', cursor=1 → cursorText='b', before='a', after=''
-    // The formatter must have received 'a' and '' but never 'b' (the cursor char).
-    const lastTwoCalls = seen.slice(-2);
-    expect(lastTwoCalls).toEqual(['a', '']);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Caret rendering (Fix B): cursor block is painted on every render — including
-// when the buffer is empty — so the user always sees where input lands.
-// ---------------------------------------------------------------------------
-
-describe('TerminalCompositor — caret always rendered while armed', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-  });
-
-  it('emits the thin-bar caret character even when buffer is empty', async () => {
-    // At end-of-buffer the compositor paints a ▏ (U+258F, LEFT ONE EIGHTH BLOCK)
-    // in the caret accent color instead of an inverse-video space block.
-    // chalk.level is forced to 1 so color SGRs are emitted; the ▏ character
-    // itself is present regardless of chalk level. Saved/restored to avoid
-    // leaking state to other suites.
-    const chalkModule = await import('chalk');
-    const priorLevel = chalkModule.default.level;
-    chalkModule.default.level = 1;
-    try {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      const frame = writes.all();
-      expect(frame).toContain('▏');
-    } finally {
-      chalkModule.default.level = priorLevel;
-    }
-  });
-
-  it('emits the thin-bar caret after the buffer goes empty again', async () => {
-    // Regression target: previously the caret was suppressed when buffer.length === 0
-    // && !queued. Verify that after typing then deleting back to empty, the caret
-    // is still painted. This exercises the post-Backspace empty-buffer render.
-    const chalkModule = await import('chalk');
-    const priorLevel = chalkModule.default.level;
-    chalkModule.default.level = 1;
-    try {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-      stdin.emit('keypress', undefined, { name: 'backspace' });
-      // Buffer is now empty (and !queued). The most recent rendered frame must
-      // still contain the ▏ thin-bar caret.
-      expect(c.getBuffer().text).toBe('');
-      expect(c.getBuffer().queued).toBe(false);
-      writes.clear();
-      // Force a repaint via setOverlay change so we observe one fresh frame
-      // generated from the empty+!queued state.
-      c.setOverlay('overlay-content');
-      const frame = writes.all();
-      expect(frame).toContain('▏');
-    } finally {
-      chalkModule.default.level = priorLevel;
-    }
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Caret blink wiring: arm() starts the ticker, each interval toggles the
-// painted caret, a keystroke snaps it back to solid, and disarm() stops it.
-// ---------------------------------------------------------------------------
-
-describe('TerminalCompositor — caret blink', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let writes: ReturnType<typeof collectWrites>;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    writes = collectWrites(stdout);
-  });
-
-  afterEach(() => {
-    vi.clearAllTimers();
-    vi.useRealTimers();
-  });
-
-  it('pulses the ▏ caret on/off on the interval and resets to solid on a keystroke', async () => {
-    vi.useFakeTimers();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      caretBlink: true,
-      caretBlinkIntervalMs: 50,
-    });
-    await c.arm();
-    // First frame (armed) shows the solid caret.
-    expect(writes.all()).toContain('▏');
-
-    // One interval later the caret blinks OFF — the repaint paints a blank cell.
-    writes.clear();
-    vi.advanceTimersByTime(50);
-    expect(writes.all()).not.toContain('▏');
-
-    // A keystroke snaps the caret back to solid (and types the char).
-    writes.clear();
-    stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-    expect(writes.all()).toContain('▏');
-    expect(c.getBuffer().text).toBe('x');
-
-    // After disarm the ticker is stopped: no further repaints fire.
-    c.disarm();
-    writes.clear();
-    vi.advanceTimersByTime(500);
-    expect(writes.all()).toBe('');
-  });
-
-  it('does not start a blink timer when caretBlink is unset (default)', async () => {
-    vi.useFakeTimers();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    expect(writes.all()).toContain('▏'); // solid caret painted at arm
-    writes.clear();
-    // No ticker → no timer-driven repaints regardless of how far time advances.
-    vi.advanceTimersByTime(5000);
-    expect(writes.all()).toBe('');
-    c.disarm();
-  });
-
-  it('coalesces the caret un-hide with the edit repaint — one frame on an off-phase keystroke', async () => {
-    vi.useFakeTimers();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      caretBlink: true,
-      caretBlinkIntervalMs: 50,
-    });
-    await c.arm();
-    // Blink into the OFF phase (caret painted away).
-    writes.clear();
-    vi.advanceTimersByTime(50);
-    expect(writes.all()).not.toContain('▏');
-
-    // A printable key in the off phase: dispatchKey → applyEdit repaints once,
-    // and that single frame already shows the now-solid caret. resetVisible()
-    // must NOT add a second frame. Before the fix this painted twice.
-    const before = c.repaintCount;
-    writes.clear();
-    stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-    expect(c.repaintCount - before).toBe(1); // exactly one frame, not two
-    expect(c.getBuffer().text).toBe('x');
-    expect(writes.all()).toContain('▏'); // solid caret in that one frame
-    c.disarm();
-  });
-
-  it('un-hides an off-phase caret with exactly one repaint on a non-painting keystroke', async () => {
-    vi.useFakeTimers();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      caretBlink: true,
-      caretBlinkIntervalMs: 50,
-    });
-    await c.arm();
-    vi.advanceTimersByTime(50); // → off phase
-    // F5 is consumed by no edit handler, so dispatchKey paints nothing; the
-    // caret-blink un-hide must still issue exactly one repaint to show it solid.
-    const before = c.repaintCount;
-    writes.clear();
-    stdin.emit('keypress', undefined, { name: 'f5' });
-    expect(c.repaintCount - before).toBe(1);
-    expect(writes.all()).toContain('▏');
-    c.disarm();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Tab applies dropdown selection (Fix C): mirrors reader.ts:769-772 behavior
-// so completion works during agent turns, not just user turns.
-// ---------------------------------------------------------------------------
-
-describe('TerminalCompositor — Tab applies dropdown selection', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    resetSlashRegistry();
-    registerSlashCommand({
-      name: '/mint',
-      summary: 'stub for tab-apply coverage',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-    registerSlashCommand({
-      name: '/diagnose',
-      summary: 'second stub',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-  });
-
-  afterEach(() => {
-    resetSlashRegistry();
-  });
-
-  it('Tab on an open slash dropdown applies the selected candidate', async () => {
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-
-    // Type '/' to open the dropdown.
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-    expect(ac.candidates.length).toBeGreaterThan(0);
-
-    // Capture the selected candidate's value (selectedIndex starts at 0).
-    const selected = ac.candidates[ac.selectedIndex];
-    expect(selected).toBeDefined();
-    const expectedValue = selected!.value;
-
-    // Tab — applies selection.
-    stdin.emit('keypress', undefined, { name: 'tab' });
-
-    // Buffer should be the selected slash command + trailing space.
-    expect(c.getBuffer().text).toBe(expectedValue + ' ');
-    // Dropdown should be closed after applying.
-    expect(ac.dropdownOpen).toBe(false);
-  });
-
-  it('Tab with no dropdown open is a no-op (does not insert literal tab)', async () => {
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-
-    // Type a non-trigger character — no dropdown.
-    stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-    expect(ac.dropdownOpen).toBe(false);
-    const beforeTab = c.getBuffer().text;
-
-    stdin.emit('keypress', undefined, { name: 'tab' });
-
-    // Buffer unchanged — Tab was swallowed.
-    expect(c.getBuffer().text).toBe(beforeTab);
-  });
-
-  it('Shift+Tab does NOT apply selection (preserves any future onShiftTab semantics)', async () => {
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), autocompleteState: ac });
-    await c.arm();
-
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-    const bufferBefore = c.getBuffer().text;
-
-    // Shift+Tab — must NOT apply selection (reader.ts uses shift-tab for thinking-mode toggle).
-    stdin.emit('keypress', undefined, { name: 'tab', shift: true });
-
-    // Buffer unchanged; dropdown still open.
-    expect(c.getBuffer().text).toBe(bufferBefore);
-    expect(ac.dropdownOpen).toBe(true);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// Enter applies dropdown selection (regression: hitting Enter on an open
-// slash-command dropdown must finalize the highlighted candidate AND fire
-// onSubmit, not submit the raw partial. Mirrors reader.ts:734-748. The
-// compositor became the exclusive TTY Enter path in Stage 3e (commit
-// 4e28e5d) but the dropdown-Enter guard was never ported — this suite
-// guards against re-introducing that gap.
-// ---------------------------------------------------------------------------
-
-describe('TerminalCompositor — Enter applies dropdown selection', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    resetSlashRegistry();
-    registerSlashCommand({
-      name: '/mint',
-      summary: 'stub for enter-apply coverage',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-    registerSlashCommand({
-      name: '/diagnose',
-      summary: 'second stub so /mi is unambiguous',
-      handler: async () => ({ kind: 'noop' as const }),
-    });
-  });
-
-  afterEach(() => {
-    resetSlashRegistry();
-  });
-
-  it('Enter on an open slash dropdown applies selection AND fires onSubmit with completed text', async () => {
-    const onSubmit = vi.fn();
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      autocompleteState: ac,
-      onSubmit,
-    });
-    await c.arm();
-    c.setInputMode('idle');
-
-    // Type '/mi' — narrows the dropdown to a single candidate (/mint).
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    stdin.emit('keypress', 'm', { name: 'm', sequence: 'm' });
-    stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-    expect(ac.dropdownOpen).toBe(true);
-    const selected = ac.candidates[ac.selectedIndex];
-    expect(selected).toBeDefined();
-    const expectedValue = selected!.value;
-    expect(expectedValue).toBe('/mint');
-
-    // Enter — must apply the dropdown selection AND submit. Before the
-    // fix this submitted '/mi' raw.
-    stdin.emit('keypress', undefined, { name: 'return' });
-
-    expect(onSubmit).toHaveBeenCalledTimes(1);
-    expect(onSubmit).toHaveBeenCalledWith({ text: expectedValue + ' ', attachments: [] });
-    // Dropdown must be closed after applying.
-    expect(ac.dropdownOpen).toBe(false);
-  });
-
-  it('Enter on an open slash dropdown with NO matching candidate suppresses submit (does not send raw partial)', async () => {
-    const onSubmit = vi.fn();
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      autocompleteState: ac,
-      onSubmit,
-    });
-    await c.arm();
-    c.setInputMode('idle');
-
-    // Open the dropdown with '/'.
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-    // Force the candidate list to empty WHILE dropdownOpen stays true —
-    // exercises the "applySelection no-op" branch (kind === 'slash' &&
-    // applied === false). Mirrors reader.ts COR-2.
-    ac.candidates = [];
-    ac.selectedIndex = 0;
-
-    stdin.emit('keypress', undefined, { name: 'return' });
-
-    // No submit fired — the raw '/' must NOT escape as a non-command.
-    expect(onSubmit).not.toHaveBeenCalled();
-  });
-
-  it('Enter on an open @file dropdown finalizes path but does NOT submit (user is mid-sentence)', async () => {
-    const onSubmit = vi.fn();
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      autocompleteState: ac,
-      onSubmit,
-    });
-    await c.arm();
-    c.setInputMode('idle');
-
-    // Forge a file-trigger dropdown directly — the real file resolver is
-    // out of scope here; we only care that `kind === 'file'` causes Enter
-    // to apply-only without submitting. This is the file-completion arm
-    // of reader.ts:734-748 (kind !== 'slash' branch).
-    stdin.emit('keypress', '@', { name: '@', sequence: '@' });
-    stdin.emit('keypress', 's', { name: 's', sequence: 's' });
-    ac.dropdownOpen = true;
-    ac.trigger = { kind: 'file', query: 's' };
-    ac.candidates = [{ value: '@src/foo.ts', summary: '' }];
-    ac.selectedIndex = 0;
-
-    stdin.emit('keypress', undefined, { name: 'return' });
-
-    // onSubmit must NOT have fired — Enter on a file completion only
-    // accepts the path; the user is still composing the prompt body.
-    expect(onSubmit).not.toHaveBeenCalled();
-    // The path should have been applied to the buffer.
-    expect(c.getBuffer().text).toContain('@src/foo.ts');
-    // Dropdown closed.
-    expect(ac.dropdownOpen).toBe(false);
-  });
-
-  it('Enter on an open --flag dropdown finalizes flag but does NOT submit', async () => {
-    const onSubmit = vi.fn();
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      autocompleteState: ac,
-      onSubmit,
-    });
-    await c.arm();
-    c.setInputMode('idle');
-
-    // Forge a flag-trigger dropdown — same shape as file-trigger; the
-    // guard's `kind !== 'slash'` branch covers both.
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    stdin.emit('keypress', 'm', { name: 'm', sequence: 'm' });
-    stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-    stdin.emit('keypress', 'n', { name: 'n', sequence: 'n' });
-    stdin.emit('keypress', 't', { name: 't', sequence: 't' });
-    stdin.emit('keypress', ' ', { name: 'space', sequence: ' ' });
-    stdin.emit('keypress', '-', { name: '-', sequence: '-' });
-    stdin.emit('keypress', '-', { name: '-', sequence: '-' });
-    ac.dropdownOpen = true;
-    ac.trigger = { kind: 'flag', command: '/mint', query: '' };
-    ac.candidates = [{ value: '--continue', summary: '' }];
-    ac.selectedIndex = 0;
-
-    stdin.emit('keypress', undefined, { name: 'return' });
-
-    // Same contract as file-trigger: apply only, no submit.
-    expect(onSubmit).not.toHaveBeenCalled();
-    expect(c.getBuffer().text).toContain('--continue');
-    expect(ac.dropdownOpen).toBe(false);
-  });
-
-  it('Enter in streaming mode on an open slash dropdown applies selection before queueing', async () => {
-    // Streaming mode (default) doesn't fire onSubmit on Enter — it sets
-    // queued=true so the parent can pick the completed buffer up at
-    // stream-end. The dropdown selection must still be applied first so
-    // the queued buffer holds the completed command, not the raw partial.
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout,
-      stdin,
-      onCancel: vi.fn(),
-      autocompleteState: ac,
-    });
-    await c.arm();
-    // No setInputMode call — default is 'streaming'.
-
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    stdin.emit('keypress', 'm', { name: 'm', sequence: 'm' });
-    stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-    expect(ac.dropdownOpen).toBe(true);
-
-    stdin.emit('keypress', undefined, { name: 'return' });
-
-    // Buffer cleared after commit (new FIFO contract); queue holds '/mint '.
-    // Dropdown is closed. The completion happened BEFORE the commit, so the
-    // payload in the FIFO is the resolved command, never the raw partial.
-    expect(c.getBuffer()).toEqual({ text: '', queued: true });
-    expect(ac.dropdownOpen).toBe(false);
-    expect(c.getPendingCount()).toBe(1);
-    // Drain to verify the queued payload is the completed command.
-    const onSubmit = vi.fn();
-    c.setOnSubmit(onSubmit);
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenCalledWith({ text: '/mint ', attachments: [] });
-  });
-});
-
-describe('TerminalCompositor — input mode + onSubmit (Stage 3a)', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-  });
-
-  describe('default mode (streaming)', () => {
-    it('starts in streaming mode (legacy default for all existing callers)', () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      expect(c.getInputMode()).toBe('streaming');
-    });
-
-    it('Enter in streaming mode commits to FIFO and clears live buffer', async () => {
-      // New contract: Enter in streaming mode commits the buffer to pendingSubmissions
-      // and clears the live input. queued=true mirrors pendingCount>0.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-    });
-
-    it('Enter in streaming mode does NOT fire onSubmit (legacy callers without idle mode)', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).not.toHaveBeenCalled();
-      expect(c.getBuffer().queued).toBe(true);
-    });
-  });
-
-  describe('idle mode + onSubmit', () => {
-    it('Enter in idle mode fires onSubmit(buffer) immediately', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'hi', attachments: [] });
-      expect(onSubmit).toHaveBeenCalledTimes(1);
-    });
-
-    it('Enter in idle mode clears the buffer + queued flag', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(c.getBuffer()).toEqual({ text: '', queued: false });
-    });
-
-    it('Enter in idle mode with no onSubmit installed falls back to streaming queue behavior', async () => {
-      // Defensive: if a caller flips to idle but never sets onSubmit,
-      // Enter must not be silently swallowed — it commits to the FIFO instead
-      // so the payload is preserved for the next readLine call.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      // Live buffer cleared; 'h' is in the FIFO.
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-    });
-
-    it('Enter on empty buffer in idle mode is a no-op (does not fire onSubmit)', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('streaming → idle transition (auto-submit queued buffer)', () => {
-    it('flushes queued buffer via onSubmit when transitioning streaming → idle', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      // Stays in default streaming mode; user commits a message mid-stream.
-      stdin.emit('keypress', 'q', { name: 'q', sequence: 'q' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      // New contract: live buffer cleared; 'q' is in the FIFO.
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-      // Stream ends, surface flips mode → onSubmit fires with the FIFO payload.
-      c.setInputMode('idle');
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'q', attachments: [] });
-      expect(c.getBuffer()).toEqual({ text: '', queued: false });
-    });
-
-    it('streaming → idle with NO queued buffer does NOT fire onSubmit', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-      // No Enter — buffer is dirty but not queued
-      c.setInputMode('idle');
-      expect(onSubmit).not.toHaveBeenCalled();
-      // The unqueued buffer survives the transition (user can keep typing in idle mode).
-      expect(c.getBuffer().text).toBe('x');
-    });
-
-    it('streaming → idle with queued buffer + NO onSubmit leaves FIFO intact (legacy contract)', async () => {
-      // With no onSubmit handler, → idle cannot drain. The FIFO payload survives
-      // so the next readLine (which installs a handler) can flush it.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', 'q', { name: 'q', sequence: 'q' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      // Live buffer is already cleared at Enter-time.
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-      c.setInputMode('idle');
-      // No handler → FIFO untouched; queued stays true.
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-    });
-
-    it('idle → streaming is a no-op (no flush)', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'q', { name: 'q', sequence: 'q' });
-      // Don't press Enter yet; transition back to streaming
-      c.setInputMode('streaming');
-      expect(onSubmit).not.toHaveBeenCalled();
-      expect(c.getBuffer().text).toBe('q');
-    });
-
-    it('idle → idle with queued buffer + handler flushes (race between readLine calls)', async () => {
-      // Scenario: between two readLine calls, the user types + Enter
-      // while no onSubmit is installed. The Enter falls through to the
-      // streaming-queue branch (sets queued=true). When the next
-      // readLine installs a handler + calls setInputMode('idle'), the
-      // widened flush invariant fires the handler immediately so the
-      // queued buffer isn't stranded waiting on a second Enter press.
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-      // No handler installed yet — type + Enter commits to FIFO and clears live buffer.
-      stdin.emit('keypress', 'r', { name: 'r', sequence: 'r' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-      // Now install handler + transition idle → idle (no-op transition
-      // but should still flush per the widened invariant).
-      const onSubmit = vi.fn();
-      c.setOnSubmit(onSubmit);
-      c.setInputMode('idle');
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'r', attachments: [] });
-      expect(c.getBuffer()).toEqual({ text: '', queued: false });
-    });
-  });
-
-  describe('setOnSubmit (post-construction installation)', () => {
-    it('installs a handler that wasn\'t set at construction', async () => {
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      const onSubmit = vi.fn();
-      c.setOnSubmit(onSubmit);
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'x', attachments: [] });
-    });
-
-    it('setOnSubmit(null) clears a previously-installed handler', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      c.setOnSubmit(null);
-      stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).not.toHaveBeenCalled();
-      // Falls back to queue (FIFO) behavior: live buffer cleared, pending count 1.
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-    });
-  });
-
-  describe('ordered-operation invariant — clear state BEFORE handler fires', () => {
-    it('idle Enter: getBuffer() observed from inside onSubmit returns the cleared state', async () => {
-      let observed: { text: string; queued: boolean } | null = null;
-      const c = new TerminalCompositor({
-        stdout, stdin, onCancel: vi.fn(),
-        onSubmit: () => { observed = c.getBuffer(); },
-      });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'r', { name: 'r', sequence: 'r' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      // The handler must see state already cleared — otherwise a reentrant
-      // call (handler queues another Enter) would double-fire on stale buffer.
-      expect(observed).toEqual({ text: '', queued: false });
-    });
-
-    it('streaming → idle flush: getBuffer() observed from inside onSubmit returns the cleared state', async () => {
-      let observed: { text: string; queued: boolean } | null = null;
-      const c = new TerminalCompositor({
-        stdout, stdin, onCancel: vi.fn(),
-        onSubmit: () => { observed = c.getBuffer(); },
-      });
-      await c.arm();
-      stdin.emit('keypress', 'q', { name: 'q', sequence: 'q' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      c.setInputMode('idle');
-      expect(observed).toEqual({ text: '', queued: false });
-    });
-  });
-});
-
-// ── Multi-message queue (FIFO) ─────────────────────────────────────────────
-//
-// Verifies the new multi-message type-ahead contract introduced alongside the
-// commit-on-Enter change: pressing Enter commits to a FIFO (pendingSubmissions)
-// and CLEARS the live input so the user can compose the NEXT message. Each
-// → idle transition drains exactly one payload (oldest first). N queued
-// messages require N sequential turns (streaming → idle cycles) to fully drain.
-describe('TerminalCompositor — multi-message queue', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-  });
-
-  it('queuing 3 messages mid-stream increments pendingCount and drains FIFO in order', async () => {
-    const onSubmit = vi.fn();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-    await c.arm();
-
-    // Helper: type text + press Enter (commits, clears live buffer).
-    const typeAndEnter = (text: string) => {
-      for (const ch of text) stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      stdin.emit('keypress', undefined, { name: 'return' });
-    };
-
-    typeAndEnter('one');
-    expect(c.getBuffer()).toEqual({ text: '', queued: true });
-    expect(c.getPendingCount()).toBe(1);
-
-    typeAndEnter('two');
-    expect(c.getBuffer()).toEqual({ text: '', queued: true });
-    expect(c.getPendingCount()).toBe(2);
-
-    typeAndEnter('three');
-    expect(c.getBuffer()).toEqual({ text: '', queued: true });
-    expect(c.getPendingCount()).toBe(3);
-
-    // Drain first payload: streaming → idle delivers 'one' (oldest).
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenCalledTimes(1);
-    expect(onSubmit).toHaveBeenLastCalledWith({ text: 'one', attachments: [] });
-    expect(c.getPendingCount()).toBe(2);
-
-    // Drain second: streaming → idle delivers 'two'.
-    c.setInputMode('streaming');
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenCalledTimes(2);
-    expect(onSubmit).toHaveBeenLastCalledWith({ text: 'two', attachments: [] });
-    expect(c.getPendingCount()).toBe(1);
-
-    // Drain third: streaming → idle delivers 'three'.
-    c.setInputMode('streaming');
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenCalledTimes(3);
-    expect(onSubmit).toHaveBeenLastCalledWith({ text: 'three', attachments: [] });
-    expect(c.getPendingCount()).toBe(0);
-    expect(c.getBuffer()).toEqual({ text: '', queued: false });
-  });
-
-  it('in-progress draft survives draining a queued message', async () => {
-    const onSubmit = vi.fn();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-    await c.arm();
-
-    // Commit one message.
-    for (const ch of 'queued') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    stdin.emit('keypress', undefined, { name: 'return' });
-    expect(c.getPendingCount()).toBe(1);
-
-    // Type a second message WITHOUT Enter — live buffer = 'draft'.
-    for (const ch of 'draft') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    expect(c.getBuffer()).toEqual({ text: 'draft', queued: true });
-    expect(c.getPendingCount()).toBe(1);
-
-    // Drain the committed 'queued' message.
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenCalledTimes(1);
-    expect(onSubmit).toHaveBeenCalledWith({ text: 'queued', attachments: [] });
-
-    // Live buffer 'draft' is intact; queue is empty.
-    expect(c.getBuffer().text).toBe('draft');
-    expect(c.getPendingCount()).toBe(0);
-    expect(c.getBuffer().queued).toBe(false);
-  });
-
-  it('render indicator shows [queued] for 1 and [N queued] for N>1', async () => {
-    const writes = collectWrites(stdout);
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-
-    // Commit first message → render should show '[queued]'.
-    writes.clear();
-    for (const ch of 'msg1') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    stdin.emit('keypress', undefined, { name: 'return' });
-    expect(writes.all()).toContain('[queued]');
-
-    // Commit second message → render should show '[2 queued]'.
-    writes.clear();
-    for (const ch of 'msg2') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    stdin.emit('keypress', undefined, { name: 'return' });
-    expect(writes.all()).toContain('[2 queued]');
-  });
-
-  it('↑ pulls the newest queued message (LIFO) for editing and re-Enter re-commits it', async () => {
-    const onSubmit = vi.fn();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-    await c.arm();
-
-    const typeAndEnter = (text: string) => {
-      for (const ch of text) stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      stdin.emit('keypress', undefined, { name: 'return' });
-    };
-
-    typeAndEnter('one');
-    typeAndEnter('two');
-    expect(c.getPendingCount()).toBe(2);
-
-    // ↑ on the (empty) live buffer pulls the NEWEST queued message ('two')
-    // back for editing; the older 'one' stays queued.
-    stdin.emit('keypress', undefined, { name: 'up' });
-    expect(c.getBuffer()).toEqual({ text: 'two', queued: true });
-    expect(c.getPendingCount()).toBe(1);
-
-    // Edit the recalled draft and re-Enter → re-commits to the BACK of the FIFO.
-    stdin.emit('keypress', '!', { name: '!', sequence: '!' });
-    stdin.emit('keypress', undefined, { name: 'return' });
-    expect(c.getBuffer()).toEqual({ text: '', queued: true });
-    expect(c.getPendingCount()).toBe(2);
-
-    // Drain: FIFO order is 'one' (oldest) then the edited 'two!'.
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenLastCalledWith({ text: 'one', attachments: [] });
-    c.setInputMode('streaming');
-    c.setInputMode('idle');
-    expect(onSubmit).toHaveBeenLastCalledWith({ text: 'two!', attachments: [] });
-    expect(c.getPendingCount()).toBe(0);
-  });
-
-  it('↑ does not pull the queue while the live buffer holds a draft', async () => {
-    // The queued-message pull is gated on an EMPTY buffer. With a draft in
-    // progress and no history wired, ↑ is a no-op — the draft and the queue
-    // are both preserved (↑ never clobbers an in-progress message).
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    for (const ch of 'queued') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    stdin.emit('keypress', undefined, { name: 'return' }); // commits 'queued', buffer → ''
-    expect(c.getPendingCount()).toBe(1);
-    for (const ch of 'draft') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-    stdin.emit('keypress', undefined, { name: 'up' }); // buffer non-empty → gate blocks the pull
-    expect(c.getBuffer()).toEqual({ text: 'draft', queued: true });
-    expect(c.getPendingCount()).toBe(1);
-  });
-});
-
-// Mock readClipboardImage so the bracketed-paste / Ctrl+V branches can be
-// exercised deterministically without spawning osascript.
-vi.mock('./input/clipboard-image.js', () => ({
-  readClipboardImage: vi.fn(),
-}));
-
-describe('TerminalCompositor — attachments + paste (Stage 3c)', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-  let mockReadClipboardImage: ReturnType<typeof vi.fn>;
-
-  beforeEach(async () => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-    // Reset the mock between tests so call counts don't leak.
-    const mod = await import('./input/clipboard-image.js');
-    mockReadClipboardImage = mod.readClipboardImage as unknown as ReturnType<typeof vi.fn>;
-    mockReadClipboardImage.mockReset();
-  });
-
-  // Helper: emit the bracketed-paste markers.
-  const startPaste = (s = stdin) => s.emit('keypress', undefined, { sequence: '\x1b[200~' });
-  const endPaste = (s = stdin) => s.emit('keypress', undefined, { sequence: '\x1b[201~' });
-
-  // Helper: build a minimal ImageAttachment-shaped object. The compositor
-  // doesn't inspect the contents — it just holds and forwards them.
-  const fakeImage = (label = 'img.png') => ({
-    kind: 'image' as const,
-    mediaType: 'image/png' as const,
-    base64: 'AAAA',
-    sourceLabel: label,
-  });
-
-  describe('bracketed paste — image-only clipboard (zero-char paste)', () => {
-    it('probes clipboard on a zero-char bracketed paste and pushes the resulting attachment', async () => {
-      const fake = fakeImage('screenshot.png');
-      mockReadClipboardImage.mockResolvedValue(fake);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      startPaste();
-      // No keystrokes inside the paste window — pure image paste.
-      endPaste();
-      // Let the osascript Promise drain.
-      await new Promise((r) => setImmediate(r));
-      expect(mockReadClipboardImage).toHaveBeenCalledTimes(1);
-      expect(c.getAttachments()).toEqual([fake]);
-    });
-
-    it('flags "[clipboard: no image found]" when zero-char paste finds nothing', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      startPaste();
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      expect(c.getAttachments()).toEqual([]);
-      // The clipboardFailureMsg is surfaced via repaint; it's hard to
-      // assert directly without scraping log-update writes, so we just
-      // assert no attachment landed. The render-path test below covers
-      // the visible status row.
-    });
-  });
-
-  describe('multi-message queue — attachment round-trip on ↑ recall', () => {
-    it('↑ restores a queued message\'s attachments when pulling it back for editing', async () => {
-      const fake = fakeImage('shot.png');
-      mockReadClipboardImage.mockResolvedValue(fake);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // Compose text + attach an image, then commit (streaming Enter). The
-      // attachment is snapshotted into the FIFO payload and the live list clears.
-      for (const ch of 'look') stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      expect(c.getAttachments()).toEqual([fake]);
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(c.getPendingCount()).toBe(1);
-      expect(c.getAttachments()).toEqual([]); // live list cleared at commit
-      // ↑ on the empty buffer pulls the message back: text AND the snapshotted
-      // attachment are restored to the live buffer for editing / re-commit.
-      stdin.emit('keypress', undefined, { name: 'up' });
-      expect(c.getBuffer().text).toBe('look');
-      expect(c.getAttachments()).toEqual([fake]);
-      expect(c.getPendingCount()).toBe(0);
-    });
-  });
-
-  describe('bracketed paste — non-empty paste (text + maybe image)', () => {
-    it('probes clipboard silently after a text paste — Finder copy attaches both', async () => {
-      const fake = fakeImage('mixed.png');
-      mockReadClipboardImage.mockResolvedValue(fake);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      startPaste();
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      expect(c.getBuffer().text).toBe('hi');
-      expect(c.getAttachments()).toEqual([fake]);
-    });
-
-    it('NO failure message surfaced after non-empty paste when clipboard has no image (silent miss)', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      startPaste();
-      stdin.emit('keypress', 't', { name: 't', sequence: 't' });
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      expect(c.getAttachments()).toEqual([]);
-      // No flag-missing marker — we'd need to scrape the rendered frame
-      // for the dim message string to assert its absence robustly.
-      // Falling back to "no attachment, no exception" as a smoke check.
-    });
-  });
-
-  describe('Ctrl+V — explicit clipboard image read', () => {
-    it('pushes a clipboard image on Ctrl+V', async () => {
-      const fake = fakeImage('ctrlv.png');
-      mockReadClipboardImage.mockResolvedValue(fake);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      expect(c.getAttachments()).toEqual([fake]);
-    });
-
-    it('guards against concurrent osascript spawns from rapid Ctrl+V', async () => {
-      // Build a never-resolving Promise so the in-flight flag stays set
-      // for the duration of the rapid presses. The .finally clears it
-      // when the Promise resolves; we drain in afterEach.
-      let resolveProbe: (v: unknown) => void;
-      const probe = new Promise((r) => { resolveProbe = r; });
-      mockReadClipboardImage.mockReturnValue(probe);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      expect(mockReadClipboardImage).toHaveBeenCalledTimes(1);
-      // Drain the probe so afterEach's mock-restore doesn't bleed.
-      resolveProbe!(null);
-      await new Promise((r) => setImmediate(r));
-    });
-  });
-
-  describe('Backspace — drops last attachment when buffer is empty', () => {
-    it('Backspace on empty buffer pops the last attachment', async () => {
-      mockReadClipboardImage.mockResolvedValue(fakeImage('a.png'));
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      expect(c.getAttachments()).toHaveLength(1);
-      stdin.emit('keypress', undefined, { name: 'backspace' });
-      expect(c.getAttachments()).toHaveLength(0);
-    });
-
-    it('Backspace on non-empty buffer edits text, does NOT touch attachments', async () => {
-      mockReadClipboardImage.mockResolvedValue(fakeImage('keep.png'));
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', undefined, { name: 'backspace' });
-      expect(c.getBuffer().text).toBe('h');
-      expect(c.getAttachments()).toHaveLength(1);
-    });
-  });
-
-  describe('Enter — attachments ride along onSubmit payload', () => {
-    it('idle Enter delivers text + attachments together', async () => {
-      const onSubmit = vi.fn();
-      mockReadClipboardImage.mockResolvedValue(fakeImage('attached.png'));
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).toHaveBeenCalledWith({
-        text: 'h',
-        attachments: [fakeImage('attached.png')],
-      });
-      // Compositor cleared both buffer AND attachments after submit.
-      expect(c.getAttachments()).toEqual([]);
-      expect(c.getBuffer().text).toBe('');
-    });
-
-    it('idle Enter on empty buffer + ≥1 attachment STILL submits (attachment-only message)', async () => {
-      const onSubmit = vi.fn();
-      mockReadClipboardImage.mockResolvedValue(fakeImage('only.png'));
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).toHaveBeenCalledWith({
-        text: '',
-        attachments: [fakeImage('only.png')],
-      });
-    });
-
-    it('streaming → idle flush also delivers attachments accumulated during the stream', async () => {
-      const onSubmit = vi.fn();
-      mockReadClipboardImage.mockResolvedValue(fakeImage('mid-stream.png'));
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      // Stay in streaming mode (default). User pastes mid-stream + presses Enter.
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      // Buffer is committed (FIFO); onSubmit hasn't fired yet.
-      expect(onSubmit).not.toHaveBeenCalled();
-      // Attachments were snapshotted into the FIFO payload at Enter-time.
-      // getAttachments() reflects the LIVE buffer's attachments — cleared to [].
-      expect(c.getAttachments()).toHaveLength(0);
-      // Stream ends — surface flips to idle. Now the queued submission flushes.
-      c.setInputMode('idle');
-      expect(onSubmit).toHaveBeenCalledWith({
-        text: 'h',
-        attachments: [fakeImage('mid-stream.png')],
-      });
-    });
-  });
-
-  describe('disarm/rearm — attachment state does not leak between sessions', () => {
-    it('resetState clears attachments + pasting flags', async () => {
-      mockReadClipboardImage.mockResolvedValue(fakeImage('orphan.png'));
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      stdin.emit('keypress', undefined, { name: 'v', ctrl: true });
-      await new Promise((r) => setImmediate(r));
-      startPaste(); // pasting = true; will be reset by disarm
-      expect(c.getAttachments()).toHaveLength(1);
-      c.disarm();
-      // resetState should have cleared attachments.
-      expect(c.getAttachments()).toEqual([]);
-    });
-  });
-
-  // Regression: multi-line clipboard paste was prematurely submitting in idle
-  // mode (or queueing in streaming mode) at the first embedded `\r` because
-  // the Enter handler treated pasted line breaks as user-submission Enter.
-  // Stage 3 ported bracketed-paste markers + the `pasting` flag, but missed
-  // the legacy reader's Enter-while-pasting → insert-literal-`\n` branch
-  // (reader.ts:721-725) and never enabled `\x1b[?2004h` in arm() so the
-  // markers themselves wouldn't fire in production. Both gaps closed
-  // together; tests below cover the full multi-line paste round-trip.
-  describe('bracketed paste — multi-line content (regression)', () => {
-    it('Enter (CR) inside a bracketed paste inserts a literal newline, NOT submits', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      startPaste();
-      // macOS-style paste: "line1\rline2" — line break sent as CR.
-      stdin.emit('keypress', 'l', { name: 'l', sequence: 'l' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', 'n', { name: 'n', sequence: 'n' });
-      stdin.emit('keypress', 'e', { name: 'e', sequence: 'e' });
-      stdin.emit('keypress', '1', { name: '1', sequence: '1' });
-      stdin.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-      stdin.emit('keypress', 'l', { name: 'l', sequence: 'l' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', 'n', { name: 'n', sequence: 'n' });
-      stdin.emit('keypress', 'e', { name: 'e', sequence: 'e' });
-      stdin.emit('keypress', '2', { name: '2', sequence: '2' });
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      // Critical invariant: onSubmit must NOT have fired mid-paste. The
-      // user's submission is what should drive that — not the terminal.
-      expect(onSubmit).not.toHaveBeenCalled();
-      // Buffer holds both lines joined by a real `\n`.
-      expect(c.getBuffer().text).toBe('line1\nline2');
-    });
-
-    it('User-typed Enter AFTER a bracketed paste ends DOES submit (idle mode)', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      startPaste();
-      stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-      stdin.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-      stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      expect(onSubmit).not.toHaveBeenCalled();
-      // Now the user explicitly hits Enter — pasting flag is false, so
-      // submit fires with the full multi-line buffer.
-      stdin.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'a\nb', attachments: [] });
-    });
-
-    it('Enter inside a bracketed paste in streaming mode does NOT set queued=true', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // Streaming mode is the default — no setInputMode call.
-      startPaste();
-      stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-      stdin.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-      stdin.emit('keypress', 'y', { name: 'y', sequence: 'y' });
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      // queued must remain false — pasted `\r` is content, not submission.
-      expect(c.getBuffer()).toEqual({ text: 'x\ny', queued: false });
-    });
-
-    it('Enter inside a bracketed paste keeps queued=true while a message is committed (mirror invariant)', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // Pre-stage: user types 'a' + Enter → commits 'a' to FIFO, live buffer → ''.
-      stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(c.getBuffer()).toEqual({ text: '', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-      // A multi-line paste arrives. The `\r` mid-paste edits the LIVE buffer but
-      // does NOT pop the FIFO — 'a' is still committed. `queued` mirrors
-      // pendingSubmissions (length 1), so it stays true; the message drains on
-      // the next → idle transition.
-      startPaste();
-      stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-      stdin.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-      stdin.emit('keypress', 'c', { name: 'c', sequence: 'c' });
-      endPaste();
-      await new Promise((r) => setImmediate(r));
-      // Live buffer holds only the pasted content ('b\nc'); 'a' is still in FIFO,
-      // so queued stays true (mirror of pendingSubmissions.length > 0).
-      expect(c.getBuffer()).toEqual({ text: 'b\nc', queued: true });
-      expect(c.getPendingCount()).toBe(1);
-    });
-  });
-
-  // Bracketed-paste truncation: when a paste exceeds the line or char
-  // threshold, the buffer keeps a compact `[Pasted text #N +M lines]`
-  // placeholder while the full content is stashed in pasteRegistry and
-  // re-expanded at submit. Best-UX trade-off — small pastes stay inline
-  // so users see what they pasted; large pastes don't blow out the input
-  // area.
-  describe('bracketed paste — large-paste truncation', () => {
-    // Helper: stream a paste burst whose body is `text`. Splits on the
-    // `\n` boundary because the dispatchKey paste path treats `\r` as
-    // a literal newline insertion (CR-shaped paste mid-burst).
-    const pasteText = (text: string, s = stdin) => {
-      startPaste(s);
-      for (const ch of text) {
-        if (ch === '\n') {
-          s.emit('keypress', '\r', { name: 'return', sequence: '\r' });
-        } else {
-          s.emit('keypress', ch, { name: ch, sequence: ch });
-        }
-      }
-      endPaste(s);
-    };
-
-    it('short paste (under both thresholds) stays inline — no placeholder', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // 3 lines = 2 newlines, below the 5-newline threshold AND below 1000 chars.
-      pasteText('alpha\nbeta\ngamma');
-      await new Promise((r) => setImmediate(r));
-      // Buffer holds the literal pasted text — no truncation.
-      expect(c.getBuffer().text).toBe('alpha\nbeta\ngamma');
-      expect(c.getBuffer().text).not.toContain('[Pasted text');
-    });
-
-    it('5+ line paste collapses into `[Pasted text #<nonce> +N lines]` placeholder', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // 6 lines = 5 newlines — at the 5-newline threshold, triggers truncation.
-      const big = 'a\nb\nc\nd\ne\nf';
-      pasteText(big);
-      await new Promise((r) => setImmediate(r));
-      // `getBuffer().text` returns the EXPANDED form (placeholder
-      // already swapped back) — that's the contract for the streaming-
-      // flush snapshot reader.
-      expect(c.getBuffer().text).toBe(big);
-      // To inspect the visible buffer we have to round-trip through
-      // submission. Use idle Enter to surface the placeholder via
-      // displayText.
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured).not.toBeNull();
-      expect(captured!.text).toBe(big);
-      expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-    });
-
-    it('single-line ≥1000-char paste collapses into `+N chars` placeholder', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      // 1500-char single-line paste — newline count is 0, so we hit
-      // the char threshold and get a `+N chars` label.
-      const big = 'x'.repeat(1500);
-      pasteText(big);
-      await new Promise((r) => setImmediate(r));
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.text).toBe(big);
-      expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+1500 chars\]$/);
-    });
-
-    it('no-truncation submission omits displayText from the payload', async () => {
-      // Existing call-sites deep-match `{ text, attachments }` — adding
-      // displayText: undefined would break them. Verify the contract
-      // explicitly: untruncated submissions have NO displayText key.
-      mockReadClipboardImage.mockResolvedValue(null);
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'hi', attachments: [] });
-    });
-
-    it('multiple pastes get distinct nonces — each placeholder has a unique 8-hex-char id', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      pasteText('a\nb\nc\nd\ne\nf');     // first paste — 6 lines
-      stdin.emit('keypress', ' ', { name: 'space', sequence: ' ' });
-      pasteText('p\nq\nr\ns\nt\nu');     // second paste — 6 lines
-      await new Promise((r) => setImmediate(r));
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.text).toBe('a\nb\nc\nd\ne\nf p\nq\nr\ns\nt\nu');
-      expect(captured!.displayText).toMatch(
-        /^\[Pasted text #[0-9a-f]{8} \+6 lines\] \[Pasted text #[0-9a-f]{8} \+6 lines\]$/,
-      );
-    });
-
-    it('single Backspace at trailing `]` atomically deletes the whole placeholder', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      pasteText('a\nb\nc\nd\ne\nf');
-      // After the paste, cursor is parked at end of placeholder. Single
-      // Backspace should kill the whole `[Pasted text #<nonce> +6 lines]`
-      // token (and drop the registry entry).
-      stdin.emit('keypress', undefined, { name: 'backspace' });
-      expect(c.getBuffer().text).toBe('');
-      // Submitting after the delete must NOT carry the stale content
-      // forward — registry entry was dropped.
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      // Type something + Enter to actually exercise the submit path.
-      stdin.emit('keypress', 'z', { name: 'z', sequence: 'z' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.text).toBe('z');
-      expect(captured!.displayText).toBeUndefined();
-    });
-
-    it('forward-delete at leading `[` atomically deletes the placeholder', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      pasteText('a\nb\nc\nd\ne\nf');
-      // Move cursor to start of placeholder via Home.
-      stdin.emit('keypress', undefined, { name: 'home' });
-      stdin.emit('keypress', undefined, { name: 'delete' });
-      expect(c.getBuffer().text).toBe('');
-    });
-
-    it('disarm/rearm clears the paste registry between sessions', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      pasteText('a\nb\nc\nd\ne\nf');
-      c.disarm();
-      await c.arm();
-      // Fresh paste after rearm gets a new nonce — registry was cleared.
-      pasteText('p\nq\nr\ns\nt\nu');
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-    });
-
-    it('submitting expanded text clears the registry — next paste gets a fresh nonce', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      pasteText('a\nb\nc\nd\ne\nf');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-      // Second paste in the same arm cycle — registry cleared by submit,
-      // next paste gets a new random nonce.
-      pasteText('p\nq\nr\ns\nt\nu');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-    });
-
-    it('streaming → idle flush also expands placeholders for the submission', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      // Stay in streaming mode (default). User pastes a big blob mid-stream,
-      // queues with Enter; the stream end's setInputMode('idle') must flush
-      // with the EXPANDED text.
-      pasteText('a\nb\nc\nd\ne\nf');
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).not.toHaveBeenCalled();
-      c.setInputMode('idle');
-      expect(onSubmit).toHaveBeenCalledWith(
-        expect.objectContaining({
-          text: 'a\nb\nc\nd\ne\nf',
-          attachments: [],
-        }),
-      );
-      const call = onSubmit.mock.calls[0]![0] as { displayText?: string };
-      expect(call.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-    });
-
-    // H1: mid-buffer paste — pre-existing text before the cursor must be
-    // preserved; only the pasted span gets collapsed into a placeholder.
-    it('mid-buffer paste: pre-existing text before cursor is preserved', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-
-      // Type a prefix so the cursor sits at position 6.
-      for (const ch of 'hello ') {
-        stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      }
-
-      // Paste a large blob (6 lines → triggers truncation at cursor=6).
-      const pasted = 'a\nb\nc\nd\ne\nf';
-      pasteText(pasted);
-      await new Promise((r) => setImmediate(r));
-
-      // getBuffer() expands placeholders → full content visible.
-      expect(c.getBuffer().text).toBe('hello ' + pasted);
-
-      // Submit: text = expanded, displayText = prefix + placeholder.
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured).not.toBeNull();
-      expect(captured!.text).toBe('hello a\nb\nc\nd\ne\nf');
-      expect(captured!.displayText).toMatch(/^hello \[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-    });
-
-    it('mid-buffer paste: text typed after the paste is also preserved', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-
-      for (const ch of 'pre ') {
-        stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      }
-      pasteText('a\nb\nc\nd\ne\nf');
-      await new Promise((r) => setImmediate(r));
-
-      // Type a suffix after the paste (cursor is at end of placeholder).
-      for (const ch of ' suf') {
-        stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      }
-
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.text).toBe('pre a\nb\nc\nd\ne\nf suf');
-      expect(captured!.displayText).toMatch(/^pre \[Pasted text #[0-9a-f]{8} \+6 lines\] suf$/);
-    });
-
-    // M1: threshold boundary values — off-by-one coverage for both dimensions.
-    describe('threshold boundary values', () => {
-      it('exactly 4 newlines (one below threshold) stays inline — no placeholder', async () => {
-        mockReadClipboardImage.mockResolvedValue(null);
-        const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-        await c.arm();
-        // 4 newlines = 5 visual lines; check is newlineCount < 5, so 4 passes through.
-        const text = 'a\nb\nc\nd\ne';
-        pasteText(text);
-        await new Promise((r) => setImmediate(r));
-        expect(c.getBuffer().text).toBe(text);
-        expect(c.getBuffer().text).not.toContain('[Pasted text');
-      });
-
-      it('exactly 5 newlines (at threshold) collapses to placeholder', async () => {
-        mockReadClipboardImage.mockResolvedValue(null);
-        const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-        await c.arm();
-        // 5 newlines = 6 visual lines; newlineCount >= 5 → truncate.
-        const text = 'a\nb\nc\nd\ne\nf';
-        pasteText(text);
-        await new Promise((r) => setImmediate(r));
-        let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-        c.setOnSubmit((p) => { captured = p as typeof captured; });
-        c.setInputMode('idle');
-        stdin.emit('keypress', undefined, { name: 'return' });
-        expect(captured!.text).toBe(text);
-        expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+6 lines\]$/);
-      });
-
-      it('exactly 999 chars (one below char threshold) stays inline — no placeholder', async () => {
-        mockReadClipboardImage.mockResolvedValue(null);
-        const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-        await c.arm();
-        // charCount < 1000 → no truncation.
-        const text = 'x'.repeat(999);
-        pasteText(text);
-        await new Promise((r) => setImmediate(r));
-        expect(c.getBuffer().text).toBe(text);
-        expect(c.getBuffer().text).not.toContain('[Pasted text');
-      });
-
-      it('exactly 1000 chars (at char threshold) collapses to +N chars placeholder', async () => {
-        mockReadClipboardImage.mockResolvedValue(null);
-        const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-        await c.arm();
-        // charCount >= 1000 → truncate.
-        const text = 'x'.repeat(1000);
-        pasteText(text);
-        await new Promise((r) => setImmediate(r));
-        let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-        c.setOnSubmit((p) => { captured = p as typeof captured; });
-        c.setInputMode('idle');
-        stdin.emit('keypress', undefined, { name: 'return' });
-        expect(captured!.text).toBe(text);
-        expect(captured!.displayText).toMatch(/^\[Pasted text #[0-9a-f]{8} \+1000 chars\]$/);
-      });
-    });
-
-    // M2: false-positive expansion safety — a user who manually types the
-    // literal placeholder format should have their text pass through unmodified.
-    it('manually-typed placeholder text passes through unexpanded when registry is empty', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-
-      // Type the literal placeholder string by hand — no actual paste.
-      const literal = '[Pasted text #1 +6 lines]';
-      for (const ch of literal) {
-        stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      }
-      stdin.emit('keypress', undefined, { name: 'return' });
-
-      // Registry is empty → expandPastePlaceholders fast-paths; the text
-      // passes through verbatim.
-      expect(onSubmit).toHaveBeenCalledWith({ text: literal, attachments: [] });
-    });
-
-    it('manually-typed placeholder with non-existent id passes through unexpanded', async () => {
-      // Registry has a random 8-hex-char nonce but user types "#99" — no
-      // hit (short hex id never matches an 8-char nonce), literal survives.
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-
-      // First, do a real paste to populate the registry with a random nonce.
-      pasteText('a\nb\nc\nd\ne\nf');
-      await new Promise((r) => setImmediate(r));
-
-      // Now also type a non-existent short hex id "#99" literal.
-      const typed = ' [Pasted text #99 +6 lines]';
-      for (const ch of typed) {
-        stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      }
-
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      stdin.emit('keypress', undefined, { name: 'return' });
-
-      // Real paste expands; typed "#99" has no registry hit, passes through.
-      expect(captured!.text).toBe('a\nb\nc\nd\ne\nf [Pasted text #99 +6 lines]');
-    });
-
-    // M3: word-delete (Option+Delete / meta+backspace) does NOT invoke the
-    // atomic-placeholder-delete path — it nibbles the placeholder word-by-word.
-    // This test documents the current behavior (not a bug, just a known gap).
-    it('meta+backspace at end of placeholder nibbles word-by-word (no atomic delete)', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      pasteText('a\nb\nc\nd\ne\nf');
-      await new Promise((r) => setImmediate(r));
-
-      // Cursor is at end of `[Pasted text #<nonce> +6 lines]`.
-      // Option+Delete fires meta+backspace — should NOT use atomic delete.
-      stdin.emit('keypress', undefined, { name: 'backspace', meta: true });
-
-      // After one meta+backspace, the buffer should still contain the opening
-      // bracket and most of the placeholder (word-delete removed the trailing
-      // `lines]` or similar chunk). It must NOT be fully empty (atomic delete
-      // would have emptied it).
-      const after = c.getBuffer().text;
-      // Registry entry still present (not deleted by word-delete path) →
-      // expansion still works on whatever fragment remains. The exact fragment
-      // depends on InputCore.deleteWordBackward word boundaries but the full
-      // paste content has NOT been expanded-and-cleared.
-      expect(after).not.toBe('');
-    });
-
-    // SEC-1: nonce collision resistance — integer-format typed ids cannot
-    // expand real paste content because real nonces are 8 hex chars.
-    it('SEC-1: typed integer-format placeholder cannot expand real paste content (nonce mismatch)', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-
-      // Do a real paste to populate the registry with one 8-hex-char nonce.
-      pasteText('a\nb\nc\nd\ne\nf');
-      await new Promise((r) => setImmediate(r));
-
-      // Type a placeholder using old-style short integer id "#1".
-      // The new regex accepts [0-9a-f]+ so "#1" still matches syntactically,
-      // but "1" is not the 8-char nonce in the registry — no expansion.
-      const typed = ' [Pasted text #1 +6 lines]';
-      for (const ch of typed) {
-        stdin.emit('keypress', ch, { name: ch, sequence: ch });
-      }
-
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      stdin.emit('keypress', undefined, { name: 'return' });
-
-      // The real paste expands; the typed "#1" has no registry hit and
-      // passes through as a literal (nonce mismatch).
-      expect(captured!.text).toBe('a\nb\nc\nd\ne\nf [Pasted text #1 +6 lines]');
-    });
-
-    // SEC-2: embedded sentinel sanitization — ensure registry content is
-    // free of bracketed-paste sentinel bytes after maybeTruncatePaste.
-    it('SEC-2: expanded paste content does not contain bracketed-paste sentinel bytes', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      c.setInputMode('idle');
-
-      // Paste a large blob that triggers truncation. The sentinel-stripping
-      // code in maybeTruncatePaste ensures the stashed content is clean.
-      pasteText('a\nb\nc\nd\ne\nf');
-      await new Promise((r) => setImmediate(r));
-
-      let captured: { text: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      stdin.emit('keypress', undefined, { name: 'return' });
-
-      // Expanded content must not contain either sentinel sequence.
-      expect(captured!.text).not.toContain('\x1b[200~');
-      expect(captured!.text).not.toContain('\x1b[201~');
-      expect(captured!.text).toBe('a\nb\nc\nd\ne\nf');
-    });
-
-    // COR-3: replaceRange-before-delete ordering — the registry entry
-    // must survive a hypothetical replaceRange throw so it is not lost.
-    // In practice InputCore.replaceRange does not throw for valid cursor
-    // positions; this test documents the happy-path sequence and verifies
-    // that after atomic placeholder delete the registry entry is gone and
-    // the buffer is cleared (both replaceRange and delete ran).
-    it('COR-3: after atomic placeholder delete, buffer is empty and registry entry is removed', async () => {
-      mockReadClipboardImage.mockResolvedValue(null);
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-      await c.arm();
-      pasteText('a\nb\nc\nd\ne\nf');
-      await new Promise((r) => setImmediate(r));
-
-      // getBuffer().text returns the EXPANDED form — paste content is present.
-      expect(c.getBuffer().text).toBe('a\nb\nc\nd\ne\nf');
-
-      // Atomic backspace: deletes the whole placeholder token (and the
-      // registry entry — replaceRange runs BEFORE delete per the F3 invariant).
-      stdin.emit('keypress', undefined, { name: 'backspace' });
-
-      // Buffer is now empty (replaceRange ran, placeholder removed).
-      expect(c.getBuffer().text).toBe('');
-
-      // Submit: expanded text equals display text (registry entry removed,
-      // no placeholder to expand) so displayText is omitted.
-      let captured: { text: string; displayText?: string; attachments: unknown[] } | null = null;
-      c.setOnSubmit((p) => { captured = p as typeof captured; });
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'z', { name: 'z', sequence: 'z' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(captured!.text).toBe('z');
-      expect(captured!.displayText).toBeUndefined();
-    });
-  });
-
-  // Stage 3b/Stage 3c parity
-  // and alt+Enter as a "soft newline" UX — explicit user intent for multi-
-  // line input without leaving the prompt. Ported to the compositor so the
-  // persistent input surface keeps the same affordance.
-  describe('shift+Enter / alt+Enter — soft newline insertion', () => {
-    it('shift+Enter inserts a literal `\\n` and does NOT submit', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-      stdin.emit('keypress', undefined, { name: 'return', shift: true });
-      stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-      expect(c.getBuffer().text).toBe('a\nb');
-      expect(onSubmit).not.toHaveBeenCalled();
-    });
-
-    it('alt+Enter (key.meta=true) inserts a literal `\\n` and does NOT submit', async () => {
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-      stdin.emit('keypress', undefined, { name: 'return', meta: true });
-      stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-      expect(c.getBuffer().text).toBe('a\nb');
-      expect(onSubmit).not.toHaveBeenCalled();
-    });
-
-    it('kitty keyboard protocol shift+Enter (`\\x1b[13;2u`) inserts a literal `\\n`', async () => {
-      // Some terminals (xterm in CSI-u mode, certain kitty configs) don't
-      // set key.shift on Enter but DO emit `\x1b[13;2u`. The compositor
-      // recognizes that sequence as shift+Enter.
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-      stdin.emit('keypress', undefined, { name: 'return', sequence: '\x1b[13;2u' });
-      stdin.emit('keypress', 'b', { name: 'b', sequence: 'b' });
-      expect(c.getBuffer().text).toBe('a\nb');
-      expect(onSubmit).not.toHaveBeenCalled();
-    });
-
-    it('plain Enter (no modifiers) still submits in idle mode', async () => {
-      // Sanity check: shift/alt are required to suppress submission.
-      // Without them, Enter resolves onSubmit as before.
-      const onSubmit = vi.fn();
-      const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-      await c.arm();
-      c.setInputMode('idle');
-      stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-      stdin.emit('keypress', undefined, { name: 'return' });
-      expect(onSubmit).toHaveBeenCalledWith({ text: 'h', attachments: [] });
-    });
-  });
-});
-
-describe('TerminalCompositor — picker mode', () => {
-  let stdout: MockStdout;
-  let stdin: MockStdin;
-
-  beforeEach(() => {
-    stdout = makeMockStdout();
-    stdin = makeMockStdin();
-  });
-
-  it('enterPickerMode flips inputMode to picker and triggers a repaint', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    const onKey = vi.fn();
-    const renderRows = vi.fn(() => ['  ? question', '  ▸ alpha', '  ↑/↓ pick · enter']);
-    c.enterPickerMode({ renderRows, onKey });
-    // renderRows is called every repaint; enterPickerMode calls repaint
-    // synchronously so renderRows fires at least once before any key.
-    expect(renderRows).toHaveBeenCalled();
-    c.exitPickerMode();
-    c.disarm();
-  });
-
-  it('enterPickerMode throws if a picker is already active', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    const ctrl = { renderRows: () => [], onKey: vi.fn() };
-    c.enterPickerMode(ctrl);
-    expect(() => c.enterPickerMode(ctrl)).toThrow(/already active/);
-    c.exitPickerMode();
-    c.disarm();
-  });
-
-  it('exitPickerMode is idempotent (no-op when no picker active)', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    expect(() => c.exitPickerMode()).not.toThrow();
-    c.exitPickerMode(); // again
-    c.disarm();
-  });
-
-  it('all keystrokes route to picker controller while picker is active', async () => {
-    const onSubmit = vi.fn();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn(), onSubmit });
-    await c.arm();
-    c.setInputMode('idle');
-    const onKey = vi.fn();
-    c.enterPickerMode({ renderRows: () => ['row'], onKey });
-    // Various keys that would normally have semantics in idle mode
-    stdin.emit('keypress', undefined, { name: 'return' });   // would normally fire onSubmit
-    stdin.emit('keypress', undefined, { name: 'up' });
-    stdin.emit('keypress', undefined, { name: 'down' });
-    stdin.emit('keypress', undefined, { name: 'escape' });
-    stdin.emit('keypress', 'a', { name: 'a', sequence: 'a' });
-    stdin.emit('keypress', ' ', { name: 'space', sequence: ' ' });
-    expect(onKey).toHaveBeenCalledTimes(6);
-    // None of those keys reached the non-picker handlers (onSubmit etc.)
-    expect(onSubmit).not.toHaveBeenCalled();
-    c.exitPickerMode();
-    c.disarm();
-  });
-
-  it('Ctrl+C in picker mode routes to onKey (not onCancel)', async () => {
-    const onCancel = vi.fn();
-    const c = new TerminalCompositor({ stdout, stdin, onCancel });
-    await c.arm();
-    const onKey = vi.fn();
-    c.enterPickerMode({ renderRows: () => [], onKey });
-    stdin.emit('keypress', undefined, { name: 'c', ctrl: true });
-    expect(onCancel).not.toHaveBeenCalled();
-    expect(onKey).toHaveBeenCalledTimes(1);
-    expect(onKey.mock.calls[0]?.[1]?.name).toBe('c');
-    expect(onKey.mock.calls[0]?.[1]?.ctrl).toBe(true);
-    c.exitPickerMode();
-    c.disarm();
-  });
-
-  it('exitPickerMode restores previous input mode', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    c.setInputMode('idle');
-    expect(c.getInputMode()).toBe('idle');
-    c.enterPickerMode({ renderRows: () => [], onKey: vi.fn() });
-    expect(c.getInputMode()).toBe('picker');
-    c.exitPickerMode();
-    expect(c.getInputMode()).toBe('idle');
-    c.disarm();
-  });
-
-  it('repaintPicker triggers a repaint when picker is active', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    const renderRows = vi.fn(() => ['row']);
-    c.enterPickerMode({ renderRows, onKey: vi.fn() });
-    renderRows.mockClear();
-    c.repaintPicker();
-    expect(renderRows).toHaveBeenCalled();
-    c.exitPickerMode();
-    c.disarm();
-  });
-
-  it('repaintPicker is a no-op when no picker is active', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    // No throw, no error — just silent return.
-    expect(() => c.repaintPicker()).not.toThrow();
-    c.disarm();
-  });
-
-  it('picker frame replaces input region (buffer text NOT rendered)', async () => {
-    // Build a compositor, type some buffer text, then enter picker mode
-    // and verify the buffer is hidden from the frame.
-    const writes = collectWrites(stdout);
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(),
-      promptTextFn: () => '> ',
-    });
-    await c.arm();
-    stdin.emit('keypress', 'h', { name: 'h', sequence: 'h' });
-    stdin.emit('keypress', 'i', { name: 'i', sequence: 'i' });
-    expect(c.getBuffer().text).toBe('hi');
-    writes.clear();
-    c.enterPickerMode({
-      renderRows: () => ['  ? PICKER_HEADER', '  ▸ option-1'],
-      onKey: vi.fn(),
-    });
-    // The picker frame's render should now contain the picker rows but
-    // NOT the user's typed buffer (the input region is rented to the picker).
-    const frameOutput = writes.all();
-    expect(frameOutput).toContain('PICKER_HEADER');
-    expect(frameOutput).toContain('option-1');
-    expect(frameOutput).not.toContain('> hi');
-    c.exitPickerMode();
-    c.disarm();
-  });
-
-  it('exit restores buffer rendering — original input survives picker turn', async () => {
-    const writes = collectWrites(stdout);
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(),
-      promptTextFn: () => '> ',
-    });
-    await c.arm();
-    stdin.emit('keypress', 'x', { name: 'x', sequence: 'x' });
-    expect(c.getBuffer().text).toBe('x');
-    c.enterPickerMode({ renderRows: () => ['picker'], onKey: vi.fn() });
-    writes.clear();
-    c.exitPickerMode();
-    // Buffer must be intact AND repainted (the input row should reappear).
-    expect(c.getBuffer().text).toBe('x');
-    expect(writes.all()).toContain('x');
-    c.disarm();
-  });
-
-  it('disarm during active picker clears picker state (defence-in-depth)', async () => {
-    const c = new TerminalCompositor({ stdout, stdin, onCancel: vi.fn() });
-    await c.arm();
-    c.enterPickerMode({ renderRows: () => ['row'], onKey: vi.fn() });
-    expect(c.getInputMode()).toBe('picker');
-    c.disarm();
-    // After disarm, input mode and picker controller are reset.
-    expect(c.getInputMode()).toBe('streaming');
-    // Re-arm and verify no picker leaked.
-    await c.arm();
-    expect(c.getInputMode()).toBe('streaming');
-    c.disarm();
-  });
-
-  it('autocomplete dropdown is reset on enterPickerMode (no bleed into picker frame)', async () => {
-    // Install an autocomplete state with an open dropdown, then enter
-    // picker mode. The dropdown rows must not appear in the picker frame.
-    const ac = createAutocompleteState();
-    const c = new TerminalCompositor({
-      stdout, stdin, onCancel: vi.fn(),
-      autocompleteState: ac,
-    });
-    await c.arm();
-    // Drive a slash so the dropdown opens. Register a fake slash so there's
-    // at least one candidate; otherwise the dropdown stays closed.
-    resetSlashRegistry();
-    registerSlashCommand({
-      name: 'fakecmd',
-      summary: 'test command',
-      handler: () => ({ handled: true, result: null }),
-    });
-    stdin.emit('keypress', '/', { name: '/', sequence: '/' });
-    expect(ac.dropdownOpen).toBe(true);
-    c.enterPickerMode({ renderRows: () => ['  ? picker'], onKey: vi.fn() });
-    expect(ac.dropdownOpen).toBe(false);
-    c.exitPickerMode();
-    c.disarm();
-    resetSlashRegistry();
-  });
-});
-
